# Only merge distinct aggregate hash tables into the global queues when full