--- a/src/processor/operator/aggregate/hash_aggregate.cpp
+++ b/src/processor/operator/aggregate/hash_aggregate.cpp
@@ -254,7 +254,7 @@ void HashAggregate::executeInternal(ExecutionContext* context) {
             break;
         }
     }
-    localState.aggregateHashTable->mergeAll();
+    localState.aggregateHashTable->mergeIfFull(0 /*tuplesToAdd*/, true /*mergeAll*/);
 }
 
 } // namespace processor