Instead of whenever their parent hash table is full. This improves performance when the distinct hash tables fill up significantly more quickly than their parents.

| Query | before | after |
| --- | --- | --- |
| `MATCH (h:hits) RETURN h.RegionID, COUNT(DISTINCT h.UserID) AS u ORDER BY u DESC LIMIT 10;` | 2.5s | 2.2s |
| `MATCH (h:hits) RETURN h.RegionID, SUM(h.AdvEngineID), COUNT(*) AS c, AVG(h.ResolutionWidth), COUNT(DISTINCT h.UserID) ORDER BY c DESC LIMIT 10;` | 2.7s | 2.3s |

In more extreme cases the difference would probably be larger.